Refactor ReportSources frontend component
Closes #13165.
