+++ components/frontend/src/report/ReportSources.jsx
@@ -1,192 +1,15 @@
-import ChecklistIcon from "@mui/icons-material/Checklist"
-import FormatListBulletedIcon from "@mui/icons-material/FormatListBulleted"
-import SettingsIcon from "@mui/icons-material/Settings"
-import { Stack, Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from "@mui/material"
-import { func, string } from "prop-types"
-import { useContext, useRef } from "react"
+import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from "@mui/material"
+import { func } from "prop-types"
+import { useContext } from "react"
 
 import { DataModelContext } from "../context/DataModel"
-import { EDIT_REPORT_PERMISSION } from "../context/Permissions"
-import { SnackbarContext } from "../context/Snackbar"
-import { reportPropType, settingsPropType, sourcePropType } from "../sharedPropTypes"
-import { SourceParameter } from "../source/SourceParameter"
-import { SourceTypeRichDescription } from "../source/SourceType"
-import { theme } from "../theme"
-import { getSourceName, getSourceTypeName, pluralize, referenceDocumentationURL } from "../utils"
+import { reportPropType, settingsPropType } from "../sharedPropTypes"
+import { getSourceName, getSourceTypeName } from "../utils"
 import { UnsortableTableHeaderCell } from "../widgets/TableHeaderCell"
 import { TableRowWithDetails } from "../widgets/TableRowWithDetails"
-import { Tabs } from "../widgets/Tabs"
 import { InfoMessage } from "../widgets/WarningMessage"
-import { metricsUsingSource, reportSources, unusedMetricTypesSupportedBySource } from "./report_utils"
-
-function reloadAfterMassEditSource(json, reload, showMessage) {
-    const nrSources = json.nr_sources_mass_edited
-    if (nrSources > 0) {
-        showMessage({
-            severity: "info",
-            title: "Mass edit",
-            description: `Changed ${nrSources} ${pluralize("source", nrSources)}`,
-        })
-    }
-    reload(json)
-}
-
-function SourceParameters({ reload, report, source, sourceUuid }) {
-    const dataModel = useContext(DataModelContext)
-    const showMessageRef = useRef(useContext(SnackbarContext))
-    const sourceType = dataModel.sources[source.type]
-    const allParameters = sourceType.parameters
-    const defaultLocationParameterKeys = ["url", "landing_url", "username", "password", "private_token"]
-    const locationParameterKeys = sourceType?.parameter_layout?.location?.parameters ?? defaultLocationParameterKeys
-    let parameters
-    if (new Set(Object.keys(allParameters)).intersection(new Set(locationParameterKeys)).size === 0) {
-        parameters = <InfoMessage title="No location parameters">This source has no location parameters.</InfoMessage>
-    } else {
-        parameters = locationParameterKeys.map((key) => (
-            <SourceParameter
-                editScope="report"
-                key={key}
-                report={report}
-                source={source}
-                sourceUuid={sourceUuid}
-                parameter={allParameters[key]}
-                parameterKey={key}
-                parameterValue={source?.parameters?.[key]}
-                reload={(json) => reloadAfterMassEditSource(json, reload, showMessageRef.current)}
-                requiredPermissions={[EDIT_REPORT_PERMISSION]}
-            />
-        ))
-    }
-    return (
-        <Stack spacing={2} margin="10px">
-            <SourceTypeRichDescription sourceTypeKey={source.type} />
-            {parameters}
-        </Stack>
-    )
-}
-SourceParameters.propTypes = {
-    reload: func,
-    report: reportPropType,
-    source: sourcePropType,
-    sourceUuid: string,
-}
-
-function Metrics({ report, source }) {
-    const dataModel = useContext(DataModelContext)
-    const metrics = metricsUsingSource(dataModel, report, source)
-    return (
-        <TableContainer>
-            <Table size="small">
-                <TableHead>
-                    <TableRow>
-                        <UnsortableTableHeaderCell label="Subject" />
-                        <UnsortableTableHeaderCell label="Metric" />
-                    </TableRow>
-                </TableHead>
-                <TableBody>
-                    {Object.entries(metrics).map(([metricUuid, metric]) => (
-                        <TableRow
-                            key={metricUuid}
-                            sx={{
-                                "&:nth-of-type(odd)": {
-                                    backgroundColor: theme.palette.action.hover,
-                                },
-                            }}
-                        >
-                            <TableCell>
-                                <a href={"#" + metric.subjectUuid}>{metric.subjectName}</a>
-                            </TableCell>
-                            <TableCell>
-                                <a href={"#" + metricUuid}>
-                                    {metric.name}
-                                    {metric.secondary_name ? " - " + metric.secondary_name : ""}
-                                </a>
-                            </TableCell>
-                        </TableRow>
-                    ))}
-                </TableBody>
-            </Table>
-        </TableContainer>
-    )
-}
-Metrics.propTypes = {
-    report: reportPropType,
-    source: sourcePropType,
-}
-
-function UnusedMetricTypes({ report, source }) {
-    const dataModel = useContext(DataModelContext)
-    const unusedMetricTypes = unusedMetricTypesSupportedBySource(dataModel, report, source)
-    if (unusedMetricTypes.size === 0) {
-        return (
-            <InfoMessage title="No unused metric types">
-                All metric types that this source supports are being used.
-            </InfoMessage>
-        )
-    }
-    return (
-        <TableContainer>
-            <Table size="small">
-                <TableHead>
-                    <TableRow>
-                        <UnsortableTableHeaderCell label="Metric type" />
-                        <UnsortableTableHeaderCell label="Description" />
-                        <UnsortableTableHeaderCell label="Why measure this metric type?" />
-                    </TableRow>
-                </TableHead>
-                <TableBody>
-                    {[...unusedMetricTypes].map((metricType) => (
-                        <TableRow
-                            key={metricType}
-                            sx={{
-                                "&:nth-of-type(odd)": {
-                                    backgroundColor: theme.palette.action.hover,
-                                },
-                            }}
-                        >
-                            <TableCell>
-                                <a href={referenceDocumentationURL(dataModel.metrics[metricType].name)}>
-                                    {dataModel.metrics[metricType].name}
-                                </a>
-                            </TableCell>
-                            <TableCell>{dataModel.metrics[metricType].description}</TableCell>
-                            <TableCell>{dataModel.metrics[metricType].rationale}</TableCell>
-                        </TableRow>
-                    ))}
-                </TableBody>
-            </Table>
-        </TableContainer>
-    )
-}
-UnusedMetricTypes.propTypes = {
-    report: reportPropType,
-    source: sourcePropType,
-}
-
-function SourceDetails({ reload, report, settings, source, sourceUuid }) {
-    const tabs = [
-        { label: "Source configuration", icon: <SettingsIcon /> },
-        { label: "Metrics using the source", icon: <ChecklistIcon /> },
-        { label: "Unused metric types", icon: <FormatListBulletedIcon /> },
-    ]
-    const panes = [
-        <SourceParameters key={sourceUuid} reload={reload} report={report} source={source} sourceUuid={sourceUuid} />,
-        <Metrics key="metrics" report={report} source={source} />,
-        <UnusedMetricTypes key="unused_metric_types" report={report} source={source} />,
-    ]
-    return (
-        <Tabs settings={settings} tabs={tabs} uuid={sourceUuid}>
-            {panes}
-        </Tabs>
-    )
-}
-SourceDetails.propTypes = {
-    reload: func,
-    report: reportPropType,
-    settings: settingsPropType,
-    source: sourcePropType,
-    sourceUuid: string,
-}
+import { reportSources, unusedMetricTypesSupportedBySource } from "./report_utils"
+import { SourceDetails } from "./SourceDetails"
 
 export function ReportSources({ reload, report, settings }) {
     const dataModel = useContext(DataModelContext)
