+++ components/frontend/src/report/ReportSources.test.jsx
@@ -1,5 +1,4 @@
 import { render, screen } from "@testing-library/react"
-import userEvent from "@testing-library/user-event"
 import history from "history/browser"
 import { vi } from "vitest"
 
@@ -10,7 +9,6 @@ import { DataModelContext } from "../context/DataModel"
 import { EDIT_REPORT_PERMISSION, PermissionsContext } from "../context/Permissions"
 import {
     asyncClickLabeledElement,
-    expectFetch,
     expectNoAccessibilityViolations,
     expectNoText,
     expectSearch,
@@ -61,8 +59,7 @@ it("shows the sources", async () => {
     expectText(/Source 2/)
 })
 
-it("shows a message for sources without url", async () => {
-    history.push("?expanded=source_uuid:0")
+it("shows sources without url", async () => {
     renderReportSources({
         subjects: {
             subject_uuid: {
@@ -77,7 +74,7 @@ it("shows a message for sources without url", async () => {
             },
         },
     })
-    expectText(/This source has no location parameters/)
+    expectText("Source type without location parameters", 2) // Source name (defaulted) and source type
 })
 
 it("counts the metrics", async () => {
@@ -172,93 +169,6 @@ it("considers a source without name the same as a source that has a name equal t
     expectText("Source type name", 2) // Source type and source name are equal
 })
 
-it("changes the value of a parameter of a source without parameter layout", async () => {
-    const showMessage = vi.fn()
-    history.push("?expanded=source_uuid:0")
-    renderReportSources(
-        {
-            subjects: {
-                subject_uuid: {
-                    metrics: {
-                        metric_uuid: {
-                            sources: {
-                                source_uuid: { type: "source_type", parameters: { url: "https://source.org" } },
-                            },
-                        },
-                    },
-                },
-            },
-        },
-        showMessage,
-    )
-    await userEvent.type(screen.getAllByLabelText(/URL/)[0], "/new{Enter}")
-    expectFetch("post", "source/source_uuid/parameter/url", { url: "https://source.org/new", edit_scope: "report" })
-    expect(showMessage).not.toHaveBeenCalled()
-})
-
-it("mass changes the value of a parameter", async () => {
-    vi.spyOn(fetchServerApi, "fetchServerApi").mockResolvedValue({ ok: true, nr_sources_mass_edited: 2 })
-    const showMessage = vi.fn()
-    history.push("?expanded=source_uuid:0")
-    renderReportSources(
-        {
-            subjects: {
-                subject_uuid: {
-                    metrics: {
-                        metric_uuid: {
-                            sources: {
-                                source_uuid: { type: "source_type", parameters: { url: "https://source.org" } },
-                                source_uuid2: { type: "source_type", parameters: { url: "https://source.org" } },
-                            },
-                        },
-                    },
-                },
-            },
-        },
-        showMessage,
-        dataModel,
-    )
-    await userEvent.type(screen.getAllByLabelText(/URL/)[0], "/new{Enter}")
-    expectFetch("post", "source/source_uuid/parameter/url", { url: "https://source.org/new", edit_scope: "report" })
-    expect(showMessage).toHaveBeenCalledWith({
-        description: "Changed 2 sources",
-        severity: "info",
-        title: "Mass edit",
-    })
-})
-
-it("changes the value of a parameter of a source with parameter layout", async () => {
-    history.push("?expanded=source_uuid:0")
-    const showMessage = vi.fn()
-    const theDataModel = { ...dataModel }
-    theDataModel.sources["source_type"].parameters = {
-        api_version: { name: "API version", type: "string" },
-    }
-    theDataModel.sources["source_type"].parameter_layout = {
-        location: { parameters: ["api_version"] },
-    }
-    renderReportSources(
-        {
-            subjects: {
-                subject_uuid: {
-                    metrics: {
-                        metric_uuid: {
-                            sources: {
-                                source_uuid: { type: "source_type", parameters: { api_version: "2" } },
-                            },
-                        },
-                    },
-                },
-            },
-        },
-        showMessage,
-        theDataModel,
-    )
-    await userEvent.type(screen.getByLabelText(/API version/), "{Backspace}3{Enter}")
-    expectFetch("post", "source/source_uuid/parameter/api_version", { api_version: "3", edit_scope: "report" })
-    expect(showMessage).not.toHaveBeenCalled()
-})
-
 function expectOrder(expected) {
     const rows = screen.getAllByText(/source1|source2/)
     for (let index = 0; index < expected.length; index++) {
@@ -312,19 +222,6 @@ it("sorts the sources by type if no names are available", async () => {
     expectOrder(["https://source1.org", "https://source2.org"]) // Source order unchanged because the source types are equal
 })
 
-it("shows the metrics using the source", async () => {
-    history.push("?expanded=source_uuid:1")
-    renderReportSources(report)
-    expectText("M1")
-})
-
-it("shows the secondary names of metrics using the source", async () => {
-    history.push("?expanded=source_uuid2:1") // Use the second metric and source for better test coverage
-    report.subjects["subject_uuid"].metrics["metric_uuid2"].secondary_name = "secondary name"
-    renderReportSources(report)
-    expectText(/M2.*secondary name/)
-})
-
 it("ignores an empty report", async () => {
     history.push("?expanded=source_uuid:1")
     renderReportSources({})
@@ -343,67 +240,8 @@ it("ignores metrics without sources", async () => {
     expectNoText("M1")
 })
 
-it("shows the unused metric types", async () => {
-    history.push("?expanded=source_uuid2:2")
-    const report = {
-        subjects: {
-            subject_uuid: {
-                metrics: {
-                    metric_uuid: {
-                        type: "metric_type",
-                        sources: {
-                            source_uuid: {
-                                type: "source_type",
-                                parameters: { url: "https://example.org" },
-                            },
-                        },
-                    },
-                    metric_uuid2: {
-                        type: "metric_type",
-                        sources: {
-                            source_uuid2: {
-                                type: "source_type_without_location_parameters",
-                                parameters: { url: "https://example.org" },
-                            },
-                        },
-                    },
-                },
-            },
-        },
-    }
-    renderReportSources(report)
-    expectText("Metric type 2")
-    const readTheDocsLink = screen.getByRole("link", { name: "Metric type 2" })
-    expect(readTheDocsLink).toHaveAttribute("href", expect.stringContaining("#metric-type-2"))
-    expectText("Metric type description")
-    expectText("Metric type rationale")
-})
-
 it("toggles a source's expanded state when its row is expanded", async () => {
     renderReportSources(report)
     await asyncClickLabeledElement(/Expand/, 0)
     expectSearch("?expanded=source_uuid%3A0")
-})
-
-it("shows a message if there are no unused metric types", async () => {
-    history.push("?expanded=source_uuid:2")
-    const report = {
-        subjects: {
-            subject_uuid: {
-                metrics: {
-                    metric_uuid: {
-                        type: "metric_type",
-                        sources: {
-                            source_uuid: {
-                                type: "source_type",
-                                parameters: { url: "https://example.org" },
-                            },
-                        },
-                    },
-                },
-            },
-        },
-    }
-    renderReportSources(report)
-    expectText("All metric types that this source supports are being used.")
 })
